create upload component
