+++ frontend/src/app/patients/ManagePatients.tsx
@@ -1,25 +1,16 @@
-import { gql, useQuery, useMutation } from "@apollo/client";
+import { gql, useQuery } from "@apollo/client";
 import React from "react";
 import { NavLink } from "react-router-dom";
 import moment from "moment";
-import { displayFullName, showError, showNotification } from "../utils";
-import { toast } from "react-toastify";
+import { displayFullName} from "../utils";
 
-// this can't be the best way to handle this?
 import {
   PATIENT_TERM_CAP,
   PATIENT_TERM_PLURAL_CAP,
 } from "../../config/constants";
 import { daysSince } from "../utils/date";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import Alert from "../commonComponents/Alert";
-
-
-const uploadPatients = gql`
-  mutation UploadPatients($patientList: Upload!) {
-    uploadPatients(patientList: $patientList)
-  }
-`;
+import PatientUpload from "./PatientUpload";
 
 const patientQuery = gql`
   query GetPatientsByFacility($facilityId: String!) {
@@ -59,8 +50,6 @@ interface Props {
 }
 
 const ManagePatients = ({ activeFacilityId, canEditUser }: Props) => {
-  const [upload] = useMutation(uploadPatients);
-
   const { data, loading, error, refetch } = useQuery<Data, {}>(patientQuery, {
     fetchPolicy: "no-cache",
     variables: {
@@ -101,32 +90,6 @@ const ManagePatients = ({ activeFacilityId, canEditUser }: Props) => {
     });
   };
 
-  const bulkUpload = async ({target: {files }}: React.ChangeEvent<HTMLInputElement>) => {
-    const fileList = files;
-    if (fileList === null) {
-      showError(toast, "Error", "File not found");
-      return;
-    } 
-    upload({variables: {patientList: fileList[0]}}).then(() => {
-      showNotification(
-        toast,
-        <Alert
-          type="success"
-          title={`Patients uploaded`}
-          body="Successfully uploaded patients."
-        />
-      );
-      refetch();
-    },
-    (_error) => {
-      showError(
-        toast,
-        `${PATIENT_TERM_CAP} Data Error`,
-        "Please check for missing data or typos."
-      );
-    });
-  }
-
   return (
     <main className="prime-home">
       <div className="grid-container">
@@ -134,7 +97,7 @@ const ManagePatients = ({ activeFacilityId, canEditUser }: Props) => {
           <div className="prime-container usa-card__container">
             <div className="usa-card__header">
               <h2> All {PATIENT_TERM_PLURAL_CAP}</h2>
-              {canEditUser ? (<>
+              {canEditUser ? (
                 <NavLink
                   className="usa-button usa-button--outline"
                   to={`/add-patient/?facility=${activeFacilityId}`}
@@ -143,12 +106,6 @@ const ManagePatients = ({ activeFacilityId, canEditUser }: Props) => {
                   <FontAwesomeIcon icon="plus" />
                   {` New ${PATIENT_TERM_CAP}`}
                 </NavLink>
-                              <input
-                              type="file"
-                              name="file"
-                              placeholder='UploadCSV...'
-                              onChange={bulkUpload}
-                            /></>
               ) : null}
 
             </div>
@@ -174,6 +131,7 @@ const ManagePatients = ({ activeFacilityId, canEditUser }: Props) => {
               )}
             </div>
           </div>
+          <PatientUpload onSuccess={refetch}/>
         </div>
       </div>
     </main>
