+++ frontend/src/app/store.ts
@@ -24,6 +24,7 @@ const initialState = {
     email: "",
     permissions: [],
     roleDescription: "",
+    isAdmin: false,
   },
 };
